+++ CHANGELOG.md
@@ -8,6 +8,12 @@ Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `
 
 ## Unreleased
 
+## 25.6.1 - 12-02-21
+
+### Fixed
+
+- SC-8613 - Added checkbox to ldap config to select to load all users during verification
+
 ## 25.6.0 - 09-02-21
 
 ### Added